Fix bugs

+++ src/rdkafka_telemetry_encode.h
@@ -68,7 +68,7 @@ typedef enum {
 } rd_kafka_telemetry_consumer_metric_name_t;
 
 typedef union {
-        int32_t intValue;
+        int64_t intValue;
         double doubleValue;
 } rd_kafka_telemetry_metric_value_t;
